requirements.txt
paho-mqtt==1.6.1
requests==2.32.3